freki/requirements.txt
apscheduler==3.6.3
flask==1.1.2
flask-restplus==0.13.0
flask-sqlalchemy==2.4.4
flask-login==0.5.0
flask-limiter==1.3.1
gunicorn==20.0.4
pefile==2019.4.18
PyJWT==1.6.4
pyminizip==0.2.4
pyopenssl==19.1.0
python-magic==0.4.18
requests==2.24.0
sqlalchemy-utils
ssdeep==3.4
werkzeug==0.16.1
yara-python==4.0.2